Safari-611.1.10/LayoutTests/crypto/subtle/aes-kw-import-raw-key-expected.txt

Test importing a raw AES-KW key with legnth 128

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-KW'
PASS key.algorithm.length is 128
PASS key.usages is ['unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
